requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
auto-face-recognition==0.0.3
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
python-restcountries==1.0.0
cycler==0.10.0
decorator==4.4.2
docopt==0.6.2
filelock==3.0.12
gast==0.3.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.35.0
gTTS==2.2.1
h5py==2.10.0
idna==2.10
JarvisAI==0.2.8
joblib==1.0.1
Js2Py==0.70
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazyme==0.0.23
librosa==0.8.0
llvmlite==0.36.0
lxml==4.6.1
Markdown==3.3.3
matplotlib==3.3.4
numpy==1.18.5
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
Pillow==8.1.0
pipwin==0.5.0
playsound==1.2.2
pooch==1.3.0
protobuf==3.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyjokes==0.6.0
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.2
pySmartDL==1.3.4
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.1
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
SoundFile==0.10.3.post1
soupsieve==2.2
SpeechRecognition==3.8.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.2
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tokenizers==0.10.1
torch==1.7.1
tqdm==4.57.0
transformers==4.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
Werkzeug==1.0.1
wikipedia==1.4.0
wrapt==1.12.1
